Move javascript into file, add button interface
Moved javascript into a seperate script.js file. Added button interfaces
in the html file, linked them in the script and added an event listener
to run a round for each button. Added divs that display both the round
result and the leaderboard. Realised that the function to run should be
placed inside a function in the event listener.